% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fido2.R
\name{fido_predict}
\alias{fido_predict}
\title{Obtain predictions from a fitted multinomial logit (ALR, CLR) Gaussian process model object and return a ts_inter object.}
\usage{
fido_predict(object, newdata, design = "~ time + P1 + V1 + subject")
}
\arguments{
\item{object}{a fitted object of class inheriting from "fido::basset".}

\item{newdata}{a ts_inter in which to look for variables with which to predict.}

\item{design}{}
}
\description{
Obtain predictions from a fitted multinomial logit (ALR, CLR) Gaussian process model object and return a ts_inter object.
}
